Do not terminate message processing early.
This fixxes waits for voluntary external agent termination.
The approach is mostly empirical - an unknown component closes
LanguageServer before exit command is processed. We need to wait more,
and the process may look like a good moment to wait for.

However, process may never terminate, so this solution is not valid and
is a temporary measure. We still need to find what causes out-of-order
stop of Language server.

Fixes #10